requirements.txt
git+https://github.com/openai/whisper.git
torchvision
torchaudio
sentencepiece
sounddevice
scipy
transformers